+++ src/main/java/view/ExampleNavigation.java
@@ -1,6 +1,7 @@
 package view;
 
 import view.components.DFrame;
+import view.pages.MainPanel;
 
 import javax.swing.*;
