+++ src/main/java/view/components/RoundedButton.java
@@ -1,4 +1,4 @@
-package view;
+package view.components;
 
 import javax.swing.*;
 import java.awt.*;
