+++ src/main/java/view/components/game/GameGUI.java
@@ -1,30 +1,30 @@
-package view;
+package view.components.game;
 
 import game.DataAccessMock;
-import game.PhotoLocation;
 import game.PhotoLocationFactory;
+import view.utils.ImageScaler;
+import view.components.RoundedButton;
 
 import javax.swing.*;
 import java.awt.*;
+import javax.imageio.ImageIO;
+import java.io.IOException;
 
-
-public class GamePanel extends JPanel {
+public class GameGUI extends JFrame {
     private JLabel roundLabel;
     private SegmentedProgressBar progressBar;
     private GameTimer gameTimer;
     private PointsDisplay pointsDisplay;
     private RoundedButton guessButton;
     private JLabel imageLabel1;
-
-    private double[] coord;
-    private InteractiveMap map =
-            new InteractiveMap(new ImageIcon(ClassLoader.getSystemResource("photos/UofTmap.jpg")),
-                    new double[]{43.66997811270511, 43.657184780883696, -79.40326917196147, -79.3848918572115});
+    private JLabel imageLabel2;
 
     private int round = 1;
     private int totalRounds = 10;
 
-    public GamePanel() {
+    public GameGUI() {
+        setTitle("Game GUI");
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         setLayout(new BorderLayout());
 
         JPanel topPanel = new JPanel(new BorderLayout());
@@ -54,24 +54,20 @@ public class GamePanel extends JPanel {
 
         JPanel imagePanel = new JPanel(new GridLayout(1, 2, 10, 0));
         imageLabel1 = new JLabel();
-
+        imageLabel2 = new JLabel();
 
         imageLabel1.setHorizontalAlignment(JLabel.CENTER);
-
+        imageLabel2.setHorizontalAlignment(JLabel.CENTER);
 
         imageLabel1.setPreferredSize(new Dimension(frameWidth, frameWidth));
-
+        imageLabel2.setPreferredSize(new Dimension(frameWidth, frameWidth));
 
         loadNewImages(frameWidth);
 
         imagePanel.add(imageLabel1);
-
-        mainCenterPanel.setLayout(new BoxLayout(mainCenterPanel, BoxLayout.X_AXIS));
+        imagePanel.add(imageLabel2);
 
         mainCenterPanel.add(imagePanel);
-
-        mainCenterPanel.add(map);
-
         mainCenterPanel.add(Box.createVerticalGlue());
 
         add(mainCenterPanel, BorderLayout.CENTER);
@@ -94,23 +90,36 @@ public class GamePanel extends JPanel {
         bottomPanel.add(progressBarPanel, BorderLayout.SOUTH);
         add(bottomPanel, BorderLayout.SOUTH);
 
+        pack();
+        setExtendedState(JFrame.MAXIMIZED_BOTH);
+        setLocationRelativeTo(null);
+        setVisible(true);
+
         gameTimer.start(this::nextRound);
     }
 
     private void loadNewImages(int frameWidth) {
         DataAccessMock dataAccess = new DataAccessMock();
         PhotoLocationFactory photoFactory = new PhotoLocationFactory(dataAccess);
-        PhotoLocation l = photoFactory.getRandomLocation();
 
-        ImageIcon fetchedImage1 = l.getPhoto();
-        coord = l.getLocation();
+        ImageIcon fetchedImage1 = photoFactory.getRandomLocation().getPhoto();
         imageLabel1.setIcon(ImageScaler.getScaledImageIcon(fetchedImage1, frameWidth, frameWidth));
+
+        try {
+            // The following line sets the RHS image. Ideally, this should be replaced
+            // with an interactive map embedding for a more engaging user experience.
+            ImageIcon mapImage = new ImageIcon(ImageIO.read(getClass().getResource("/photos/UofTmap.jpg")));
+            imageLabel2.setIcon(ImageScaler.getScaledImageIcon(mapImage, frameWidth, frameWidth));
+        } catch (IOException e) {
+            e.printStackTrace();
+            imageLabel2.setText("Map not found");
+        }
     }
 
 
+
     private void handleGuess(int frameWidth) {
-        map.setTarget(coord);
-        boolean win = (map.getDistance() < 100);
+        boolean win = Math.random() > 0.5;
         boolean barFilled = progressBar.updateRound(win);
 
         if (!barFilled) {
@@ -134,6 +143,6 @@ public class GamePanel extends JPanel {
     }
 
     public static void main(String[] args) {
-        new GamePanel();
+        new GameGUI();
     }
 }
