+++ src/main/java/view/pages/GamePanel.java
@@ -1,27 +1,36 @@
-package view;
+package view.pages;
 
 import game.DataAccessMock;
+import game.PhotoLocation;
 import game.PhotoLocationFactory;
+import view.components.game.GameTimer;
+import view.components.game.InteractiveMap;
+import view.components.game.PointsDisplay;
+import view.utils.ImageScaler;
+import view.components.RoundedButton;
+import view.components.game.SegmentedProgressBar;
+
 import javax.swing.*;
 import java.awt.*;
-import javax.imageio.ImageIO;
-import java.io.IOException;
 
-public class GameGUI extends JFrame {
+
+public class GamePanel extends JPanel {
     private JLabel roundLabel;
     private SegmentedProgressBar progressBar;
     private GameTimer gameTimer;
     private PointsDisplay pointsDisplay;
     private RoundedButton guessButton;
     private JLabel imageLabel1;
-    private JLabel imageLabel2;
+
+    private double[] coord;
+    private InteractiveMap map =
+            new InteractiveMap(new ImageIcon(ClassLoader.getSystemResource("photos/UofTmap.jpg")),
+                    new double[]{43.66997811270511, 43.657184780883696, -79.40326917196147, -79.3848918572115});
 
     private int round = 1;
     private int totalRounds = 10;
 
-    public GameGUI() {
-        setTitle("Game GUI");
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+    public GamePanel() {
         setLayout(new BorderLayout());
 
         JPanel topPanel = new JPanel(new BorderLayout());
@@ -51,20 +60,24 @@ public class GameGUI extends JFrame {
 
         JPanel imagePanel = new JPanel(new GridLayout(1, 2, 10, 0));
         imageLabel1 = new JLabel();
-        imageLabel2 = new JLabel();
+
 
         imageLabel1.setHorizontalAlignment(JLabel.CENTER);
-        imageLabel2.setHorizontalAlignment(JLabel.CENTER);
+
 
         imageLabel1.setPreferredSize(new Dimension(frameWidth, frameWidth));
-        imageLabel2.setPreferredSize(new Dimension(frameWidth, frameWidth));
+
 
         loadNewImages(frameWidth);
 
         imagePanel.add(imageLabel1);
-        imagePanel.add(imageLabel2);
+
+        mainCenterPanel.setLayout(new BoxLayout(mainCenterPanel, BoxLayout.X_AXIS));
 
         mainCenterPanel.add(imagePanel);
+
+        mainCenterPanel.add(map);
+
         mainCenterPanel.add(Box.createVerticalGlue());
 
         add(mainCenterPanel, BorderLayout.CENTER);
@@ -87,36 +100,23 @@ public class GameGUI extends JFrame {
         bottomPanel.add(progressBarPanel, BorderLayout.SOUTH);
         add(bottomPanel, BorderLayout.SOUTH);
 
-        pack();
-        setExtendedState(JFrame.MAXIMIZED_BOTH);
-        setLocationRelativeTo(null);
-        setVisible(true);
-
         gameTimer.start(this::nextRound);
     }
 
     private void loadNewImages(int frameWidth) {
         DataAccessMock dataAccess = new DataAccessMock();
         PhotoLocationFactory photoFactory = new PhotoLocationFactory(dataAccess);
+        PhotoLocation l = photoFactory.getRandomLocation();
 
-        ImageIcon fetchedImage1 = photoFactory.getRandomLocation().getPhoto();
+        ImageIcon fetchedImage1 = l.getPhoto();
+        coord = l.getLocation();
         imageLabel1.setIcon(ImageScaler.getScaledImageIcon(fetchedImage1, frameWidth, frameWidth));
-
-        try {
-            // The following line sets the RHS image. Ideally, this should be replaced
-            // with an interactive map embedding for a more engaging user experience.
-            ImageIcon mapImage = new ImageIcon(ImageIO.read(getClass().getResource("/photos/UofTmap.jpg")));
-            imageLabel2.setIcon(ImageScaler.getScaledImageIcon(mapImage, frameWidth, frameWidth));
-        } catch (IOException e) {
-            e.printStackTrace();
-            imageLabel2.setText("Map not found");
-        }
     }
 
 
-
     private void handleGuess(int frameWidth) {
-        boolean win = Math.random() > 0.5;
+        map.setTarget(coord);
+        boolean win = (map.getDistance() < 100);
         boolean barFilled = progressBar.updateRound(win);
 
         if (!barFilled) {
@@ -140,6 +140,6 @@ public class GameGUI extends JFrame {
     }
 
     public static void main(String[] args) {
-        new GameGUI();
+        new GamePanel();
     }
 }
