+++ src/main/java/view/utils/HTMLTextBuilder.java
@@ -1,4 +1,4 @@
-package view;
+package view.utils;
 
 public class HTMLTextBuilder {
